add swagger

+++ src/services/symptom-logs.service.ts
@@ -3,22 +3,39 @@ import { User } from '@interfaces/users.interface';
 import { checkIfConflict, checkIfEmpty } from './common.service';
 import { SymptomLog } from '@interfaces/symptom-logs.interface';
 import { SymptomLogEntity } from '@entity/symptom-logs.entity';
-import { CreateSymptomLogDto } from '@dtos/symptom-logs.dto';
+import {
+  CreateSymptomLogDto,
+  UpdateSymptomLogDto,
+} from '@dtos/symptom-logs.dto';
 import { UserEntity } from '@entity/users.entity';
 import { IntensityLogEntity } from '@entity/intensity-logs.entity';
+import { Paginate } from '../interfaces/internal/paginate.interface';
+import { SymptomEntity } from '../entity/symptoms.entity';
+import { CreateIntensityLogDto } from '../dtos/intensity-logs.dto';
+import IntensityLogService from './intensity-logs.service';
 
 class SymptomLogService {
-  public symptoms = SymptomLogEntity;
+  public symptomLogs = SymptomLogEntity;
+  public intensityLogsService = new IntensityLogService();
 
   public async getAllSymptomLogs(
     userId: number,
-  ): Promise<Partial<SymptomLog[]>> {
-    const symptomLogRepository = getRepository(this.symptoms);
+  ): Promise<Paginate<Partial<SymptomLog>>> {
+    const symptomLogRepository = getRepository(this.symptomLogs);
     const symptoms: SymptomLog[] = await symptomLogRepository.find({
-      select: ['id', 'intensityLogs', 'date'],
+      select: ['id', 'date'],
       where: { userId },
+      relations: ['intensityLogs'],
     });
-    return symptoms;
+    const data = symptoms.map(symptom => {
+      symptom.intensityLogs.forEach(intensityLog => {
+        delete intensityLog.createdAt;
+        delete intensityLog.updatedAt;
+      });
+    });
+
+    const total = await symptomLogRepository.count({ where: { userId } });
+    return { data: symptoms, total };
   }
 
   public async findSymptomLogById(
@@ -28,7 +45,7 @@ class SymptomLogService {
     checkIfEmpty(symptomLogId);
     checkIfEmpty(userId);
 
-    const symptomLogRepository = getRepository(this.symptoms);
+    const symptomLogRepository = getRepository(this.symptomLogs);
     const symptomLog: SymptomLog = await symptomLogRepository.findOne({
       select: ['id', 'intensityLogs', 'date'],
       where: { userId, id: symptomLogId },
@@ -47,42 +64,42 @@ class SymptomLogService {
 
     const symptom = new SymptomLogEntity();
     symptom.date = new Date(symptomData.date);
-    const intensityLogs = await this.getIntensityLogsByIds(
+    const intensityLogs = await this.createIntensityLogs(
       symptomData.intensityLogs,
     );
     symptom.intensityLogs = intensityLogs;
     const user = await this.getUserById(userId);
     symptom.user = user;
 
-    const symptomLogRepository = getRepository(this.symptoms);
+    const symptomLogRepository = getRepository(this.symptomLogs);
     await symptomLogRepository.save(symptom);
   }
 
   public async updateSymptom(
-    symptomData: Partial<CreateSymptomLogDto> & { id: number },
+    symptomData: UpdateSymptomLogDto,
     userId: number,
   ): Promise<void> {
     checkIfEmpty(symptomData);
     checkIfEmpty(userId);
-    checkIfEmpty(symptomData.id);
 
-    const symptomLogRepository = getRepository(this.symptoms);
+    const symptomLogRepository = getRepository(this.symptomLogs);
     const symptomLog: SymptomLog = await symptomLogRepository.findOne({
-      where: { id: symptomData.id },
+      where: { id: symptomData.id, userId },
+      relations: ['intensityLogs'],
     });
     checkIfConflict(!symptomLog);
-    checkIfConflict(symptomLog.user.id !== userId, 'User id does not match');
 
-    if (symptomData.date) {
-      symptomLog.date = new Date(symptomData.date);
-    }
-    if (symptomData.intensityLogs) {
-      const intensityLogs = await this.getIntensityLogsByIds(
-        symptomData.intensityLogs,
-      );
-      symptomLog.intensityLogs = intensityLogs;
-    }
-    await symptomLogRepository.update(symptomData.id, { ...symptomLog });
+    // TODO: remove unused intensity logs
+    const intensityLogs = await Promise.all(
+      symptomData.intensityLogs.map(async intensityLog => {
+        return intensityLog.id
+          ? await this.intensityLogsService.updateIntensityLog(intensityLog)
+          : await this.intensityLogsService.createIntensityLog(intensityLog);
+      }),
+    );
+    symptomLog.intensityLogs = intensityLogs;
+    symptomLog.date = new Date(symptomData.date);
+    await symptomLogRepository.save(symptomLog);
   }
 
   public async deleteSymptomLog(
@@ -92,12 +109,12 @@ class SymptomLogService {
     checkIfEmpty(symptomLogId);
     checkIfEmpty(userId);
 
-    const symptomLogRepository = getRepository(this.symptoms);
+    const symptomLogRepository = getRepository(this.symptomLogs);
     const symptomLog: SymptomLog = await symptomLogRepository.findOne({
       where: { userId, id: symptomLogId },
     });
     checkIfConflict(!symptomLog);
-    await symptomLogRepository.delete(symptomLogId);
+    await symptomLogRepository.delete({ id: symptomLogId, userId });
   }
 
   protected async getUserById(userId: number): Promise<User> {
@@ -118,6 +135,16 @@ class SymptomLogService {
     const intensityLogs = await intensityLogRepository.findByIds(ingredientIds);
     checkIfConflict(!intensityLogs);
     return intensityLogs;
+  }
+
+  protected async createIntensityLogs(
+    intensityLogs: CreateIntensityLogDto[],
+  ): Promise<IntensityLogEntity[]> {
+    return Promise.all(
+      intensityLogs.map(async intensityLog => {
+        return this.intensityLogsService.createIntensityLog(intensityLog);
+      }),
+    );
   }
 }
